requirements.txt
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.3.0
PyJWT==1.7.1
pytz==2019.1
shortuuid==0.5.0
sqlparse==0.3.0
